pyproject.toml
[tool.poetry]
name = "spycatagency"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.6"
djangorestframework = "^3.16.1"
requests = "^2.32.5"
pycountry = "^24.6.1"
django-countries = "^7.6.1"
django-filter = "^25.1"
drf-spectacular = "^0.28.0"
python-dotenv = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
